Mostly have the tool drawing and working. Tool position is updated by UpdateFixPosition.  Just working on getting the color of the section now.

+++ backend/sectionproperties.h
@@ -21,16 +21,16 @@ class SectionProperties : public QObject
 public:
     explicit SectionProperties(QObject *parent = nullptr);
 
-    SIMPLE_BINDABLE_PROPERTY(double, leftPosition)
-    SIMPLE_BINDABLE_PROPERTY(double, rightPosition)
+    SIMPLE_BINDABLE_PROPERTY(float, leftPosition)
+    SIMPLE_BINDABLE_PROPERTY(float, rightPosition)
     SIMPLE_BINDABLE_PROPERTY(QColor, color)
 
 signals:
     void sectionChanged();
 
 private:
-    Q_OBJECT_BINDABLE_PROPERTY_WITH_ARGS(SectionProperties, double, m_leftPosition, 0.0, &SectionProperties::leftPositionChanged)
-    Q_OBJECT_BINDABLE_PROPERTY_WITH_ARGS(SectionProperties, double, m_rightPosition, 0.0, &SectionProperties::rightPositionChanged)
+    Q_OBJECT_BINDABLE_PROPERTY_WITH_ARGS(SectionProperties, float, m_leftPosition, 0.0, &SectionProperties::leftPositionChanged)
+    Q_OBJECT_BINDABLE_PROPERTY_WITH_ARGS(SectionProperties, float, m_rightPosition, 0.0, &SectionProperties::rightPositionChanged)
     Q_OBJECT_BINDABLE_PROPERTY_WITH_ARGS(SectionProperties, QColor, m_color, QColor::fromRgbF(0,0,0), &SectionProperties::colorChanged)
 };
